pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "SimpleGVI"
version = "0.1.0"
description = "A simple tool for calculating Green View Index (GVI)"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "torch>=1.13.0",
    "torchvision>=0.14.0",
    "transformers>=4.29.2",
    "pillow>=9.0.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "tqdm>=4.62.0",
    "scikit-learn>=1.0.0",
    "pandas-stubs",
    "types-tqdm",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/SimpleGVI"
"Bug Tracker" = "https://github.com/yourusername/SimpleGVI/issues"

[tool.setuptools]
packages = ["modules"]
